Extract DripperHTMLLayoutTemplateStage to layout_template.py
Follows SemanticDedup pattern: each major stage in its own file.
stage.py now contains only shared utilities + bindings.
layout_template.py owns the layout clustering + propagation logic.

  layout_template.py: 2356 lines (layout clustering/propagation stage)
  stage.py: 489 lines (was 2721 — reduced by 2232 lines)

All syntax checks pass; ruff clean.
